requirements.txt
mysql-connector-python
requests
python-dotenv
flask